+++ gadgets/src/algorithms/snark/groth16.rs
@@ -6,7 +6,7 @@ use snarkos_models::{
         algorithms::snark::SNARKVerifierGadget,
         curves::{GroupGadget, PairingGadget},
         r1cs::{ConstraintSynthesizer, ConstraintSystem},
-        utilities::{alloc::AllocGadget, eq::EqGadget, uint::UInt8, ToBitsGadget, ToBytesGadget},
+        utilities::{alloc::AllocGadget, eq::EqGadget, uint8::UInt8, ToBitsGadget, ToBytesGadget},
     },
 };
